+++ packages/video-clone-engine/src/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+  ...require('./constants'),
+  ...require('./errors'),
+  ...require('./clone-report'),
+  ...require('./similarity'),
+  ...require('./stage-executor'),
+  ...require('./pipeline'),
+};